+++ bundletree/src/main/java/org/mqnaas/bundletree/Activator.java
@@ -1,0 +1,15 @@
+package org.mqnaas.bundletree;
+
+import org.osgi.framework.BundleActivator;
+import org.osgi.framework.BundleContext;
+
+public class Activator implements BundleActivator {
+
+	public void start(BundleContext context) throws Exception {
+	}
+
+	public void stop(BundleContext context) throws Exception {
+
+	}
+
+}
