+++ pom.xml
@@ -72,6 +72,11 @@
 			</dependency>
 			<dependency>
 				<groupId>org.mqnaas</groupId>
+				<artifactId>bundletree</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>org.mqnaas</groupId>
 				<artifactId>core</artifactId>
 				<version>${project.version}</version>
 			</dependency>
@@ -203,6 +208,7 @@
 	<modules>
 		<module>core.api</module>
 		<module>core</module>
+		<module>bundletree</module>
 		<module>clientprovider</module>
 	</modules>
